Fix bs4 warning for not specifying parser.
This adds a configuration option to set the default parser for
BeautifulSoup to use when creating soups. It also adds arguments for any
function/method that leads to a soup being created to be able to
override the default parser for a particular call.

This change is fully backward compatible with the current upstream
version. If none of the options/arguments added here ever get invoked,
then it will by absolute default use the 'html.parser' parser, which is
the batteries-included option and will suffice to shut up the
UserWarning that BeautifulSoup gives about not specifying a parser.